+++ src/page/page_layout/mixed_module/Mixed23.vue
@@ -208,6 +208,9 @@ export default {
         ...mapMutations({}),
         addShuffleLayout(squareIndex) {
             return (layoutItemType) => {
+                if (!this.$authority.isHasAuthority('content:layoutBlock:add')) {
+                    return;
+                }
                 this.squareIndex = squareIndex;
                 this.layoutItemType = layoutItemType;
                 switch (squareIndex) {
@@ -256,6 +259,9 @@ export default {
             };
         },
         addLayout(type) {
+            if (!this.$authority.isHasAuthority('content:layoutBlock:add')) {
+                return;
+            }
             let {navbarId} = this.$route.params;
             let id = _.get(this.activeLayout, `${this.index}.id`);
             switch (type) {
@@ -279,11 +285,17 @@ export default {
             }
         },
         editHandler() {
+            if (!this.$authority.isHasAuthority('content:layoutBlock:put')) {
+                return;
+            }
             let {navbarId} = this.$route.params;
             let id = _.get(this.activeLayout, `${this.index}.id`);
             this.$router.push({ name: 'ShuffleModule', params: {navbarId, index: this.index, operator: 'edit'}, query: {id} });
         },
         deleteHandler() {
+            if (!this.$authority.isHasAuthority('content:layoutBlock:put')) {
+                return;
+            }
             let id = _.get(this.activeLayout, `${this.index}.id`);
             this.$util.deleteLayoutItemHandler(id);
         }
