+++ src/page/page_layout/mixed_module/Mixed33.vue
@@ -216,6 +216,9 @@ export default {
             deleteLayoutDataByIndex: 'pageLayout/deleteLayoutDataByIndex'
         }),
         addShuffleLayout(command, squareIndex) {
+            if (!this.$authority.isHasAuthority('content:layoutBlock:add')) {
+                return;
+            }
             this.squareIndex = squareIndex;
             if (command === 'LINK') {
                 this.$refs.selectLinkDialog.showDialog(command);
@@ -224,6 +227,9 @@ export default {
             }
         },
         addLayout(type) {
+            if (!this.$authority.isHasAuthority('content:layoutBlock:add')) {
+                return;
+            }
             let {navbarId} = this.$route.params;
             let id = _.get(this.activeLayout, `${this.index}.id`);
             switch (type) {
@@ -247,11 +253,17 @@ export default {
             }
         },
         editHandler() {
+            if (!this.$authority.isHasAuthority('content:layoutBlock:put')) {
+                return;
+            }
             let {navbarId} = this.$route.params;
             let id = _.get(this.activeLayout, `${this.index}.id`);
             this.$router.push({ name: 'ShuffleModule', params: {navbarId, index: this.index, operator: 'edit'}, query: {id} });
         },
         deleteHandler() {
+            if (!this.$authority.isHasAuthority('content:layoutBlock:put')) {
+                return;
+            }
             let id = _.get(this.activeLayout, `${this.index}.id`);
             this.$util.deleteLayoutItemHandler(id);
         }
