+++ src/page/page_layout/mixed_module/Mixed4.vue
@@ -193,6 +193,9 @@ export default {
     methods: {
         ...mapMutations({}),
         addShuffleLayout(squareIndex) {
+            if (!this.$authority.isHasAuthority('content:layoutBlock:add')) {
+                return;
+            }
             return (layoutItemType) => {
                 this.squareIndex = squareIndex;
                 this.layoutItemType = layoutItemType;
@@ -229,6 +232,9 @@ export default {
             };
         },
         addLayout(type) {
+            if (!this.$authority.isHasAuthority('content:layoutBlock:add')) {
+                return;
+            }
             let {navbarId} = this.$route.params;
             let id = _.get(this.activeLayout, `${this.index}.id`);
             switch (type) {
@@ -252,11 +258,17 @@ export default {
             }
         },
         editHandler() {
+            if (!this.$authority.isHasAuthority('content:layoutBlock:put')) {
+                return;
+            }
             let {navbarId} = this.$route.params;
             let id = _.get(this.activeLayout, `${this.index}.id`);
             this.$router.push({ name: 'ShuffleModule', params: {navbarId, index: this.index, operator: 'edit'}, query: {id} });
         },
         deleteHandler() {
+            if (!this.$authority.isHasAuthority('content:layoutBlock:put')) {
+                return;
+            }
             let id = _.get(this.activeLayout, `${this.index}.id`);
             this.$util.deleteLayoutItemHandler(id);
         }
