+++ src/page/page_layout/mixed_module/Mixeds6.vue
@@ -189,6 +189,9 @@ export default {
         }),
         addShuffleLayout(squareIndex) {
             return (layoutItemType) => {
+                if (!this.$authority.isHasAuthority('content:layoutBlock:add')) {
+                    return;
+                }
                 this.squareIndex = squareIndex;
                 this.layoutItemType = layoutItemType;
                 switch (layoutItemType) {
@@ -217,6 +220,9 @@ export default {
             };
         },
         addLayout(type) {
+            if (!this.$authority.isHasAuthority('content:layoutBlock:add')) {
+                return;
+            }
             let {navbarId} = this.$route.params;
             let id = _.get(this.activeLayout, `${this.index}.id`);
             switch (type) {
@@ -240,11 +246,17 @@ export default {
             }
         },
         editHandler() {
+            if (!this.$authority.isHasAuthority('content:layoutBlock:put')) {
+                return;
+            }
             let {navbarId} = this.$route.params;
             let id = _.get(this.activeLayout, `${this.index}.id`);
             this.$router.push({ name: 'ShuffleModule', params: {navbarId, index: this.index, operator: 'edit'}, query: {id} });
         },
         deleteHandler() {
+            if (!this.$authority.isHasAuthority('content:layoutBlock:put')) {
+                return;
+            }
             let id = _.get(this.activeLayout, `${this.index}.id`);
             this.$util.deleteLayoutItemHandler(id);
         }
