+++ src/page/page_layout/programme_subject_module/PsTemplate4.vue
@@ -149,6 +149,9 @@ export default {
         ...mapMutations({}),
         addShuffleLayout(squareIndex) {
             return (layoutItemType) => {
+                if (!this.$authority.isHasAuthority('content:layoutBlock:add')) {
+                    return;
+                }
                 this.squareIndex = squareIndex;
                 this.layoutItemType = layoutItemType;
                 switch (layoutItemType) {
@@ -164,6 +167,9 @@ export default {
             };
         },
         addLayout(type) {
+            if (!this.$authority.isHasAuthority('content:layoutBlock:add')) {
+                return;
+            }
             let {navbarId} = this.$route.params;
             let id = _.get(this.activeLayout, `${this.index}.id`);
             switch (type) {
@@ -187,10 +193,16 @@ export default {
             }
         },
         editHandler() {
+            if (!this.$authority.isHasAuthority('content:layoutBlock:put')) {
+                return;
+            }
             let id = _.get(this.activeLayout, `${this.index}.id`);
             this.$router.push({ name: 'ProgrammeSubjectModule', params: {navbarId: this.navbarId, index: this.index, operator: 'edit'}, query: {id} });
         },
         deleteHandler() {
+            if (!this.$authority.isHasAuthority('content:layoutBlock:put')) {
+                return;
+            }
             let id = _.get(this.activeLayout, `${this.index}.id`);
             this.$util.deleteLayoutItemHandler(id);
         }
